+++ README.md
@@ -1,28 +1,33 @@
 # Sql Implementations
+
 ## Things to do understand before doing it
 - How Db write data on disk
-- How CPU process/read/write data from disk, memory, I am going to mess up with byte, binary, file system. (no idea)
+- How CPU process/read/write data from disk, memory
+- Working with byte, binary, file system. (no idea)
 - How client send a Statement to database (networking?)
 - How Statement are transformed from SQL language to things that DB can understand and execute.
 
 ## Page implementations fundamentals
-- Slotted Page 
+
+- Slotted Page
 - BTree
 - FSM
 - Hash Index
-- LMSTree ? - choose this approach or slotted page
+- LMSTree?
+
 ## Compression
 -
 -
+
 ## Join
 -
+
 ## ACID implementations
 - WAL?
 - Transactions?
 - Isolation Level implementations?
 - Locks & Latches
 ## Backup/Trigger
-- ?
 
 ## Things to do next
 - No idea, before learning more
